Dev complete, all crud operations functional

+++ controllers/contactController.js
@@ -2,9 +2,9 @@ const asyncHandler = require('express-async-handler');
 const Contact = require('../models/contactModel');
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access Public
+//@access Private
 const getContacts=asyncHandler (async (req, res) => {
-    const contact=await Contact.find();
+    const contact=await Contact.find({user_id:req.user.id});
     res.status(200).json({message:'Get all contacts',
         data:contact
     });
@@ -12,7 +12,7 @@ const getContacts=asyncHandler (async (req, res) => {
 
 //@desc Get contact by ID
 //@route GET /api/contacts/:id
-//@access Public
+//@access Private
 const getContactById=asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     // Check if the contact was found
@@ -20,12 +20,15 @@ const getContactById=asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
-    
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User not authorized to view other user's contact");
+    }
     res.status(200).json(contact); });
 
 //@desc Create a new contact
 //@route POST /api/contacts
-//@access Public
+//@access Private
 const createContact=asyncHandler(async(req, res) => {
     console.log("req.body", req.body);
     const {name,phone,email}= req.body;
@@ -35,6 +38,7 @@ const createContact=asyncHandler(async(req, res) => {
         //return res.status(400).json({ error: 'All fields are mandatory' });
     }
     const contact= await Contact.create({
+        user_id:req.user.id,
         name,
         phone,
         email
@@ -46,13 +50,18 @@ const createContact=asyncHandler(async(req, res) => {
 
 //@desc Update a contact
 //@route PUT /api/contacts/:id
-//@access Public
+//@access Private
 const updateContact=asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     // Check if the contact was found
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User not authorized to update other user's contact");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -64,13 +73,17 @@ const updateContact=asyncHandler(async (req, res) => {
 
 //@desc Delete a contact
 //@route DELETE /api/contacts/:id
-//@access Public
+//@access Private
 const deleteContact=asyncHandler(async(req, res) => {
     const contact = await Contact.findById(req.params.id);
     // Check if the contact was found
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User not authorized to delete other user's contact");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact); });
